requirements.txt
# Core dependencies
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<2.0.0
scipy>=1.7.0,<2.0.0

# Optimization
cvxpy>=1.2.0,<2.0.0
cvxopt>=1.2.0,<2.0.0
ecos>=2.0.7
scs>=3.0.0

# Machine Learning
scikit-learn==1.3.2
lightgbm==4.1.0
boruta==0.4.3
hmmlearn==0.3.0
arch==6.3.0
torch>=2.0.0
transformers>=4.25.0,<5.0.0

# Data sources - FMP Premium only
aiohttp>=3.8.0,<4.0.0
python-dotenv>=0.19.0,<1.0.0
requests>=2.28.0,<3.0.0
asyncio-throttle>=1.0.0,<2.0.0
pytrends>=4.7.3,<5.0.0

# Technical Analysis
ta>=0.10.1,<1.0.0

# Backtesting & Performance
empyrical>=0.5.5,<1.0.0
pytickersymbols>=1.0.18,<2.0.0

# Visualization
matplotlib>=3.5.0,<4.0.0
seaborn>=0.11.0,<1.0.0
plotly>=5.3.0,<6.0.0
dataframe-image>=0.1.1
lxml>=4.9.1

# Development tools
black>=22.0.0,<23.0.0
flake8>=4.0.0,<5.0.0
mypy>=0.950,<1.0.0
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.18.0,<1.0.0
pytest-cov>=3.0.0,<4.0.0
tqdm>=4.62.0,<5.0.0

# Jupyter
jupyter>=1.0.0
markdown-pdf>=1.1.1
tabulate>=0.9.0
ipykernel>=6.4.0
nbformat>=5.7.0

# Documentation
sphinx>=4.5.0,<5.0.0
sphinx-rtd-theme>=1.0.0,<2.0.0
